+++ hutool-core/src/main/java/cn/hutool/core/io/resource/UrlResource.java
@@ -3,7 +3,6 @@ package cn.hutool.core.io.resource;
 import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.io.IORuntimeException;
 import cn.hutool.core.io.IoUtil;
-import cn.hutool.core.util.CharsetUtil;
 import cn.hutool.core.util.ObjectUtil;
 import cn.hutool.core.util.URLUtil;
 
@@ -94,11 +93,6 @@ public class UrlResource implements Resource, Serializable{
 		} finally {
 			IoUtil.close(reader);
 		}
-	}
-	
-	@Override
-	public String readUtf8Str() throws IORuntimeException{
-		return readStr(CharsetUtil.CHARSET_UTF_8);
 	}
 	
 	@Override